+++ crates/matrix-sdk-ui/src/notification_client.rs
@@ -52,7 +52,7 @@ use tracing::{debug, info, instrument, trace, warn};
 
 use crate::{
     DEFAULT_SANITIZER_MODE,
-    encryption_sync_service::{EncryptionSyncMode, EncryptionSyncPermit, EncryptionSyncService},
+    encryption_sync_service::{EncryptionSyncPermit, EncryptionSyncService},
     sync_service::SyncService,
 };
 
@@ -325,7 +325,6 @@ impl NotificationClient {
         let encryption_sync = EncryptionSyncService::new_for_notifications(
             self.client.clone(),
             Some((Duration::from_secs(3), Duration::from_secs(4))),
-            EncryptionSyncMode::Notification,
         )
         .await;